MIMWU/BACKED/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
mysql-connector-python==9.0.0
pandas==2.2.2
scikit-learn==1.5.0
Werkzeug==3.0.3
